+++ .github/workflows/cmake.yml
@@ -1,6 +1,11 @@
 name: CMake Build
 
 env:
+  REALM_ATLAS_ENDPOINT: ${{ secrets.ATLAS_QA_BASE_URL }}
+  REALM_BAAS_ENDPOINT: ${{ secrets.APPSERVICES_QA_BASE_URL }}
+  REALM_ATLAS_API_KEY: ${{ secrets.ATLAS_QA_API_KEY }}
+  REALM_ATLAS_PRIVATE_API_KEY: ${{ secrets.ATLAS_QA_PRIVATE_API_KEY }}
+  REALM_ATLAS_PROJECT_ID: ${{ secrets.ATLAS_QA_PROJECT_ID}}
   REALM_CI: true
 
 on:
@@ -10,6 +15,37 @@ on:
   pull_request:
 
 jobs:
+  deploy-cluster:
+    runs-on: ubuntu-latest
+    name: Deploy Cluster
+    outputs:
+      clusterName: ${{ steps.deploy-cluster.outputs.clusterName }}
+    steps:
+      - uses: realm/ci-actions/mdb-realm/deploy@30bd9ead57756016f2e2ac7630433cf0883d1bae
+        id: deploy-cluster
+        with:
+          atlasUrl: ${{ env.REALM_ATLAS_ENDPOINT }}
+          realmUrl: ${{ env.REALM_BAAS_ENDPOINT }}
+          projectId: ${{ env.REALM_ATLAS_PROJECT_ID }}
+          apiKey: ${{ env.REALM_ATLAS_API_KEY }}
+          privateApiKey: ${{ env.REALM_ATLAS_PRIVATE_API_KEY }}
+
+  delete-cluster:
+    runs-on: ubuntu-latest
+    name: Delete Cluster
+    if: always()
+    needs:
+      - build-macos
+      - build-linux
+    steps:
+      - uses: realm/ci-actions/mdb-realm/cleanup@30bd9ead57756016f2e2ac7630433cf0883d1bae
+        with:
+          atlasUrl: ${{ env.REALM_ATLAS_ENDPOINT }}
+          realmUrl: ${{ env.REALM_BAAS_ENDPOINT }}
+          projectId: ${{ env.REALM_ATLAS_PROJECT_ID }}
+          apiKey: ${{ env.REALM_ATLAS_API_KEY }}
+          privateApiKey: ${{ env.REALM_ATLAS_PRIVATE_API_KEY }}
+
   build-macos:
     runs-on: macos-12
     name: macOS ${{ matrix.configuration }} (preset ${{ matrix.preset }}, Xcode ${{ matrix.xcode }})
@@ -26,7 +62,11 @@ jobs:
           - Debug
           - Release
         exclude:
-          - configuration: ${{ github.event_name == 'pull_request' && 'Debug' || 'none' }}
+          - configuration: ${{ (github.event_name == 'pull_request' && 'Release') || 'none' }}
+    needs:
+      - deploy-cluster
+    env:
+      REALM_ATLAS_CLUSTER_NAME: ${{ needs.deploy-cluster.outputs.clusterName }}
     steps:
       - name: Checkout
         uses: actions/checkout@v2
@@ -60,6 +100,18 @@ jobs:
         with:
           timeout-minutes: 15
 
+      - name: Test
+        working-directory: .build/cmake-preset-${{ matrix.preset }}
+        run: ./${{ matrix.configuration }}/cpprealm_exe_tests
+
+      - name: Publish Test Report
+        uses: mikepenz/action-junit-report@v3
+        if: always()
+        with:
+          report_paths: '.build/**/TestResults.xml'
+          annotate_only: true
+          require_tests: true
+
   build-linux:
     runs-on: ubuntu-22.04
     name: Linux ${{ matrix.configuration }} (${{ matrix.compiler.name }} ${{ matrix.compiler.version }})
@@ -77,8 +129,11 @@ jobs:
           - Debug
           - Release
         exclude:
-          - configuration: ${{ github.event_name == 'pull_request' && 'Debug' || 'none' }}
-
+          - configuration: ${{ (github.event_name == 'pull_request' && 'Release') || 'none' }}
+    needs:
+      - deploy-cluster
+    env:
+      REALM_ATLAS_CLUSTER_NAME: ${{ needs.deploy-cluster.outputs.clusterName }}
     steps:
       - name: Checkout
         uses: actions/checkout@v2
@@ -140,3 +195,15 @@ jobs:
         uses: mxschmitt/action-tmate@v3
         with:
           timeout-minutes: 15
+
+      - name: Test
+        working-directory: .build/cmake-preset-linux
+        run: ./${{ matrix.configuration }}/cpprealm_exe_tests
+
+      - name: Publish Test Report
+        uses: mikepenz/action-junit-report@v3
+        if: always()
+        with:
+          report_paths: '.build/**/TestResults.xml'
+          annotate_only: true
+          require_tests: true
