+++ tests/app_tests.cpp
@@ -16,10 +16,9 @@ TEST_CASE("app", "[app]") {
         CHECK(*synced_realm.object<AllTypesObject>(1)._id == 1);
     }
 
-#if TARGET_OS_SIMULATOR == 0 && TARGET_OS_IPHONE == 0 && TARGET_OS_WATCH == 0 && TARGET_OS_TV == 0
     SECTION("login_username_password") {
         auto app_id = Admin::shared().create_app();
-        auto app = realm::App(app_id, "http://localhost:9090");
+        auto app = realm::App(app_id, Admin::shared().base_url());
         CHECK_THROWS(app.login(realm::App::Credentials::apple("id_token")).get_future().get());
 
         app.register_user("foo@mongodb.com", "foobar").get_future().get();
@@ -27,5 +26,4 @@ TEST_CASE("app", "[app]") {
                 realm::App::Credentials::username_password("foo@mongodb.com", "foobar")).get_future().get();
         CHECK(!user.access_token().empty());
     }
-#endif
 }
